+++ DynaLock.Framework/IBuilder.cs
@@ -1,0 +1,7 @@
+﻿namespace DynaLock.Framework
+{
+    public interface IBuilder<T>
+    {
+        T Build();
+    }
+}